Refactor

+++ source/rust/shared/src/util.rs
@@ -1,7 +1,7 @@
 use nalgebra::Vec2;
 
 
-pub fn is_point_on_line(p: Vec2<f32>, l1: Vec2<f32>, l2: Vec2<f32>) -> bool {
+pub fn is_point_on_line(p: Vec2<f64>, l1: Vec2<f64>, l2: Vec2<f64>) -> bool {
 	let distance_from_line =
 		((l2.y - l1.y) * p.x - (l2.x - l1.x) * p.y + l2.x * l1.y - l1.x * l2.y)
 		.abs()
